facereco_multiprocessing/requirements.txt
Pillow==5.1.0
face_recognition==1.2.2
Flask==1.0.2
gevent==1.3.1
numpy==1.14.3
pymysql==0.8.1
